# services/marketing-ops/openapi/marketing-ops.v1.yaml
openapi: 3.1.0
info:
  title: Nexus Marketing Ops API
  version: 1.0.0
  description: REST contract for the operational campaign workspace.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /campaigns/{id}/lead-sources: {"get":{"operationId":"listLeadSources","summary":"listLeadSources","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"listLeadSources","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}},"post":{"operationId":"createLeadSource","summary":"createLeadSource","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeadSourceInput"}}}},"responses":{"201":{"description":"createLeadSource","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/lead-sources/{sourceId}: {"patch":{"operationId":"updateLeadSource","summary":"updateLeadSource","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"sourceId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"},{"$ref":"#/components/parameters/IfMatch"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeadSourcePatch"}}}},"responses":{"200":{"description":"updateLeadSource","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/leads: {"get":{"operationId":"listCampaignLeads","summary":"listCampaignLeads","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"listCampaignLeads","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/lead-imports/preview: {"post":{"operationId":"previewLeadImport","summary":"previewLeadImport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["sourceId","rows"],"properties":{"sourceId":{"type":"string","format":"uuid"},"rows":{"type":"array","minItems":1,"maxItems":500,"items":{"$ref":"#/components/schemas/LeadImportRow"}}}}}}},"responses":{"201":{"description":"previewLeadImport","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/lead-imports/{previewId}: {"get":{"operationId":"getLeadImportPreview","summary":"getLeadImportPreview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"previewId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"getLeadImportPreview","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/lead-imports/{previewId}/confirm: {"post":{"operationId":"confirmLeadImport","summary":"confirmLeadImport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"previewId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"required":["decisions"],"properties":{"decisions":{"type":"array","maxItems":500,"items":{"type":"object","additionalProperties":false,"required":["rowIndex","action"],"properties":{"rowIndex":{"type":"integer","minimum":0,"maximum":499},"action":{"enum":["create","link","skip"]},"contactId":{"type":"string","format":"uuid"}}}}}}}}},"responses":{"200":{"description":"confirmLeadImport","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/result-reports: {"get":{"operationId":"listResultReports","summary":"listResultReports","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"listResultReports","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}},"post":{"operationId":"createResultReport","summary":"createResultReport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultReportInput"}}}},"responses":{"201":{"description":"createResultReport","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/result-reports/{reportId}: {"patch":{"operationId":"updateResultReport","summary":"updateResultReport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"reportId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"},{"$ref":"#/components/parameters/IfMatch"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultReportInput"}}}},"responses":{"200":{"description":"updateResultReport","headers":{"ETag":{"$ref":"#/components/headers/ETag"}},"content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/result-reports/{reportId}/revisions: {"get":{"operationId":"listResultReportRevisions","summary":"listResultReportRevisions","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"reportId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"listResultReportRevisions","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/lead-capture-reviews: {"get":{"operationId":"listLeadCaptureReviews","summary":"listLeadCaptureReviews","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"listLeadCaptureReviews","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /campaigns/{id}/lead-capture-reviews/{captureId}/resolve: {"post":{"operationId":"resolveLeadCaptureReview","summary":"resolveLeadCaptureReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"captureId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"$ref":"#/components/parameters/IdempotencyKey"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeadCaptureDecision"}}}},"responses":{"200":{"description":"resolveLeadCaptureReview","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /results: {"get":{"operationId":"getLeadResults","summary":"getLeadResults","parameters":[],"responses":{"200":{"description":"getLeadResults","content":{"application/json":{"schema":{"type":"object","required":["data"],"properties":{"data":{}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /public/capture/{publicId}: {"options":{"operationId":"preflightPublicLeadCapture","summary":"preflightPublicLeadCapture","parameters":[{"name":"publicId","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{32}$"}}],"servers":[{"url":"/"}],"security":[],"responses":{"204":{"description":"preflightPublicLeadCapture"},"default":{"$ref":"#/components/responses/Error"}}},"post":{"operationId":"capturePublicLead","summary":"capturePublicLead","parameters":[{"name":"publicId","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{32}$"}}],"servers":[{"url":"/"}],"security":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicLeadCapture"}}}},"responses":{"202":{"description":"capturePublicLead","content":{"application/json":{"schema":{"type":"object","required":["accepted"],"properties":{"accepted":{"const":true}}}}}},"default":{"$ref":"#/components/responses/Error"}}}}
  /public/capture/{publicId}/whatsapp: {"get":{"operationId":"redirectPublicWhatsApp","summary":"redirectPublicWhatsApp","parameters":[{"name":"publicId","in":"path","required":true,"schema":{"type":"string","pattern":"^[A-Za-z0-9_-]{32}$"}}],"servers":[{"url":"/"}],"security":[],"responses":{"302":{"description":"redirectPublicWhatsApp"},"default":{"$ref":"#/components/responses/Error"}}}}
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  headers:
    ETag:
      description: Quoted current resource or aggregate version.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
  parameters:
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^(W/)?("[1-9][0-9]*"|[1-9][0-9]*)$' } }
  responses:
    Error:
      description: Stable public error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    LeadSourceInput: {"type":"object","additionalProperties":false,"required":["name","channel","kind"],"properties":{"name":{"type":"string","minLength":1,"maxLength":160},"channel":{"enum":["meta_ads","google_ads","linkedin_ads","email","whatsapp","organic","google_maps","other"]},"kind":{"enum":["landing_page","whatsapp","manual"]},"classification":{"enum":["cold","lead"]},"actionId":{"type":["string","null"],"format":"uuid"},"externalAccountId":{"type":["string","null"],"maxLength":200},"externalCampaignId":{"type":["string","null"],"maxLength":200},"allowedOrigins":{"type":"array","maxItems":20,"items":{"type":"string","format":"uri"}},"whatsappPhone":{"type":["string","null"]},"enabled":{"type":"boolean"}},"description":"landing_page requires exact allowedOrigins; whatsapp requires destination phone; google_maps defaults cold. Attribution fields are immutable."}
    LeadSourcePatch: {"type":"object","additionalProperties":false,"minProperties":1,"properties":{"name":{"type":"string","minLength":1,"maxLength":160},"allowedOrigins":{"type":"array","maxItems":20,"items":{"type":"string","format":"uri"}},"whatsappPhone":{"type":["string","null"]},"enabled":{"type":"boolean"}}}
    LeadImportRow: {"type":"object","additionalProperties":false,"required":["name","occurredAt"],"properties":{"name":{"type":"string","minLength":1,"maxLength":200},"email":{"type":"string","format":"email"},"phone":{"type":"string","maxLength":40},"company":{"type":"string","maxLength":200},"externalId":{"type":"string","maxLength":200},"occurredAt":{"type":"string","format":"date-time"},"classification":{"enum":["cold","lead"]}},"anyOf":[{"required":["email"]},{"required":["phone"]}]}
    ResultReportInput: {"type":"object","additionalProperties":false,"required":["sourceId","periodFrom","periodTo","timeZone","metrics"],"properties":{"sourceId":{"type":"string","format":"uuid"},"actionId":{"type":["string","null"],"format":"uuid"},"periodFrom":{"type":"string","format":"date"},"periodTo":{"type":"string","format":"date"},"timeZone":{"type":"string"},"metrics":{"type":"object","additionalProperties":false,"minProperties":1,"properties":{"sent":{"type":"integer","minimum":0},"delivered":{"type":"integer","minimum":0},"opened":{"type":"integer","minimum":0},"clicked":{"type":"integer","minimum":0},"responded":{"type":"integer","minimum":0},"qualified":{"type":"integer","minimum":0},"sales":{"type":"integer","minimum":0},"spend":{"type":"number","minimum":0},"revenue":{"type":"number","minimum":0}}},"notes":{"type":["string","null"],"maxLength":4000}}}
    LeadCaptureDecision: {"type":"object","additionalProperties":false,"required":["action"],"properties":{"action":{"enum":["create","link","skip"]},"contactId":{"type":"string","format":"uuid"}}}
    PublicLeadCapture: {"type":"object","additionalProperties":false,"required":["submissionId","name"],"properties":{"submissionId":{"type":"string","minLength":8,"maxLength":128},"name":{"type":"string","minLength":1,"maxLength":200},"email":{"type":"string","format":"email"},"phone":{"type":"string","maxLength":40},"company":{"type":"string","maxLength":200},"website":{"type":"string","maxLength":200},"utm":{"type":"object","additionalProperties":false,"properties":{"source":{"type":"string","maxLength":200},"medium":{"type":"string","maxLength":200},"campaign":{"type":"string","maxLength":200},"content":{"type":"string","maxLength":200},"term":{"type":"string","maxLength":200}}}},"anyOf":[{"required":["email"]},{"required":["phone"]}]}
...
    PublicErrorCode:
      type: string
      enum:
        - validation_error
        - unauthorized
        - auth_unavailable
        - tenant_required
        - forbidden
        - not_found
        - origin_forbidden
        - feature_disabled
        - idempotency_key_required
        - precondition_required
        - payload_too_large
        - idempotency_conflict
        - idempotency_in_progress
        - approval_conflict
        - approval_expired
        - approval_target_changed
        - self_approval_forbidden
        - invalid_supersession
        - version_conflict
        - invalid_transition
        - campaign_requirements_missing
        - campaign_read_only
        - primary_owner_required
        - participant_role_invalid
        - participant_exists
        - participant_not_found
        - reference_not_verified
        - artifact_not_owned
        - artifact_not_found
        - artifact_compensation_failed
        - material_type_not_allowed
        - material_too_large
        - material_empty
        - material_exists
        - material_not_found
        - version_conflict
        - invalid_item_transition
        - item_requirements_missing
        - item_content_required
        - item_blocked
        - item_terminal
        - assignee_not_authorized
        - dependency_exists
        - dependency_not_found
        - dependency_self
        - dependency_cycle
        - dependency_campaign_mismatch
        - dependency_terminal
        - artifact_exists
        - artifact_asset_mismatch
        - artifact_link_not_found
        - content_hash_mismatch
        - dependency_unavailable
        - dependency_invalid_response
        - rate_limited
        - internal_error
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId]
          properties:
            code: { $ref: '#/components/schemas/PublicErrorCode' }
            message: { type: string }
            correlationId: { type: string, minLength: 1, maxLength: 128 }
            details: {}
...